model/requirements.txt
absl-py==2.1.0
astunparse==1.6.3
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
cycler==0.11.0
fonttools==4.38.0
gast==0.3.3
google-auth==2.31.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.62.2
h5py==2.10.0
idna==3.7
importlib-metadata==6.7.0
joblib==1.3.2
kafka-python==2.0.2
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keras-vggface==0.6
kiwisolver==1.4.5
Markdown==3.4.4
MarkupSafe==2.1.5
matplotlib==3.5.3
mtcnn==0.1.1
numpy==1.18.5
oauthlib==3.2.2
opencv-python==4.10.0.84
opt-einsum==3.3.0
packaging==24.0
pandas==1.3.5
Pillow==9.5.0
protobuf==3.20.3
py4j==0.10.9.7
pyarrow==12.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.2
pyspark==3.4.3
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.0.2
scipy==1.4.1
six==1.16.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==2.3.0
threadpoolctl==3.1.0
typing_extensions==4.7.1
urllib3==2.0.7
Werkzeug==2.2.3
wrapt==1.16.0
zipp==3.15.0
